pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "hardmode"
version = "0.1.0"
description = "Raspberry Pi 5 AI device — Claude LLM, SH1106 OLED, AI Camera (IMX500), AI HAT+ (Hailo-8L), I2S mic"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.40.0",
    "python-dotenv>=1.0.0",
    "numpy>=1.26.0",
    "Pillow>=10.0.0",
    # OLED display — Inland 1.3" SH1106 128x64 over I2C
    "luma.oled>=3.13.0",
    "luma.core>=2.4.2",
    # Audio — I2S MEMS microphone via sounddevice/ALSA
    "sounddevice>=0.4.6",
    "scipy>=1.11.0",
    # NOTE: picamera2 and hailort must be installed via system packages on the Pi.
    # See scripts/setup.sh for instructions.
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.23.0",
    "black>=24.0.0",
    "ruff>=0.3.0",
]

[project.scripts]
hardmode = "src.main:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*", "config*"]

[tool.black]
line-length = 88
target-version = ["py311"]

[tool.ruff]
line-length = 88
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
